refactor(components/table): pass controller via params
