+++ packages/components/table/table-widget/xm-table-widget.component.html
@@ -3,7 +3,10 @@
               xmTableDirective
               xmTableSelectionDirective
               [xmTableSelectionConfig]="config"
-              [xmTableDirectiveConfig]="config">
+              [xmTableDirectiveConfig]="config"
+              [xmTableDirectiveController]="this.collectionControllerResolver.get() | async"
+
+>
 
   <mat-card *ngIf="xmTableDirective.context$ | async as context" class="no-horizontal-paddings mb-3">
